Cleaned up nearly the entire project
