+++ Console.cpp
@@ -1,2 +1,3 @@
 #include "Console.h"
 std::deque<ConsoleMessage> Console::messages;
+bool Console::isOpen = true;
