+++ SceneManager.h
@@ -2,22 +2,14 @@
 #include "Scene.h"
 
 class SceneManager {
+private:
+	static std::map<std::string, Scene> scenes;
+
 public:
-	static std::unordered_map<std::string, Scene> scenes;
 	static Scene* activeScene;
-	static void AddScene(const std::string& path) {
-		scenes[path] = Scene();
-	}
-	static void OnEntityDestroyed(unsigned int entity) {
-		activeScene->OnDestroyEntity(entity);
-	}
-	static bool IsSceneLoaded(const std::string& path) {
-		return scenes.find(path) != scenes.end();
-	}
-	static void ChangeScene(const std::string& path) {
-		activeScene = &scenes[path];
-	}
-	static void AddEntity(unsigned int entity) {
-		activeScene->entities.push_back(entity);
-	}
+	static void AddScene(const std::string& path);
+	static void OnEntityDestroyed(unsigned int entity);
+	static bool IsSceneLoaded(const std::string& path);
+	static void ChangeScene(const std::string& path);
+	static void AddEntity(unsigned int entity);
 };
