+++ View.h
@@ -1,12 +1,7 @@
 #pragma once
 #include <glm/glm.hpp>
-#include <glm/ext/matrix_transform.hpp>
-#include <glm/ext/matrix_clip_space.hpp>
 #include "Camera.h"
-#include "ShaderManager.h"
-#include "Time.h"
 #include "FrameBuffer.h"
-#include "InputManager.h"
 class View {
 public:
 	glm::vec2 position;
@@ -26,8 +21,5 @@ public:
 		this->dimensions = dimensions;
 		this->camera = camera;
 		this->hasCamera = true;
-	}
-	void Update(bool inRuntime) {
-		glViewport(position.x, position.y, dimensions.x, dimensions.y);
 	}
 };
