+++ imgui.ini
@@ -13,7 +13,7 @@ DockId=0x00000006,0
 Pos=314,19
 Size=657,431
 Collapsed=0
-DockId=0x00000009,0
+DockId=0x00000005,0
 
 [Window][Components]
 Pos=992,19
@@ -92,31 +92,31 @@ DockId=0x00000004,0
 Pos=314,19
 Size=657,431
 Collapsed=0
-DockId=0x00000009,0
+DockId=0x00000005,0
 
 [Window][Game]
 Pos=314,19
 Size=657,431
 Collapsed=0
-DockId=0x00000009,1
+DockId=0x00000005,1
 
 [Window][Viewport]
 Pos=314,19
 Size=676,483
 Collapsed=0
-DockId=0x00000009,0
+DockId=0x00000005,0
 
 [Window][Scene View]
 Pos=313,19
 Size=677,483
 Collapsed=0
-DockId=0x00000009,0
+DockId=0x00000005,0
 
 [Window][Game View]
-Pos=668,19
-Size=322,483
+Pos=313,19
+Size=677,483
 Collapsed=0
-DockId=0x0000000A,0
+DockId=0x00000005,1
 
 [Window][Stacked 1]
 Pos=435,274
@@ -124,8 +124,8 @@ Size=409,172
 Collapsed=0
 
 [Window][Adding Script]
-Pos=498,310
-Size=284,100
+Pos=512,321
+Size=256,77
 Collapsed=0
 
 [Table][0xD0F0C6E3,2]
@@ -133,15 +133,13 @@ Column 0  Weight=1.0000
 Column 1  Weight=1.0000
 
 [Docking][Data]
-DockSpace         ID=0x33675C32 Window=0x5B816B74 Pos=0,19 Size=1280,701 Split=X
-  DockNode        ID=0x00000001 Parent=0x33675C32 SizeRef=311,701 Split=Y Selected=0xD04A4B96
-    DockNode      ID=0x00000003 Parent=0x00000001 SizeRef=311,304 Selected=0xA628E342
-    DockNode      ID=0x00000004 Parent=0x00000001 SizeRef=311,395 Selected=0xD04A4B96
-  DockNode        ID=0x00000002 Parent=0x33675C32 SizeRef=967,701 Split=X
-    DockNode      ID=0x00000007 Parent=0x00000002 SizeRef=677,701 Split=Y
-      DockNode    ID=0x00000005 Parent=0x00000007 SizeRef=676,483 Split=X Selected=0x08AA43A5
-        DockNode  ID=0x00000009 Parent=0x00000005 SizeRef=353,483 CentralNode=1 Selected=0x08AA43A5
-        DockNode  ID=0x0000000A Parent=0x00000005 SizeRef=322,483 Selected=0x642CEEBB
-      DockNode    ID=0x00000006 Parent=0x00000007 SizeRef=676,216 Selected=0x49278EEE
-    DockNode      ID=0x00000008 Parent=0x00000002 SizeRef=288,701 Selected=0xFCB30AB3
+DockSpace       ID=0x33675C32 Window=0x5B816B74 Pos=0,19 Size=1280,701 Split=X
+  DockNode      ID=0x00000001 Parent=0x33675C32 SizeRef=311,701 Split=Y Selected=0xD04A4B96
+    DockNode    ID=0x00000003 Parent=0x00000001 SizeRef=311,304 Selected=0xA628E342
+    DockNode    ID=0x00000004 Parent=0x00000001 SizeRef=311,395 Selected=0xD04A4B96
+  DockNode      ID=0x00000002 Parent=0x33675C32 SizeRef=967,701 Split=X
+    DockNode    ID=0x00000007 Parent=0x00000002 SizeRef=677,701 Split=Y
+      DockNode  ID=0x00000005 Parent=0x00000007 SizeRef=676,483 CentralNode=1 Selected=0x08AA43A5
+      DockNode  ID=0x00000006 Parent=0x00000007 SizeRef=676,216 Selected=0x49278EEE
+    DockNode    ID=0x00000008 Parent=0x00000002 SizeRef=288,701 Selected=0xFCB30AB3
